# api/openapi.yaml
openapi: 3.0.3
info:
  title: arca-dns Controller API
  description: |
    REST API for arca-dns controller - BGP Anycast + ECMP DNS system with split control/data plane.

    This API provides zone management in two modes:
    - **JSON Mode**: Structured JSON for programmatic access
    - **Raw Mode**: Accept raw BIND zone files for migration/compatibility

    Key features:
    - DNSSEC signing with central key management
    - Zone versioning with ETag support for optimistic concurrency control
    - Pluggable storage backends (SQLite, PostgreSQL, MySQL, Git, etcd)
    - Artifact distribution for agents with integrity verification
    - Health/readiness/status endpoints are served on the API listener (default port 8080); Prometheus metrics are served on the observability listener (default port 9053)
  version: 1.0.0
  contact:
    name: arca-dns
    url: https://github.com/akam1o/arca-dns
  license:
    name: See LICENSE file
    url: https://github.com/akam1o/arca-dns/blob/main/LICENSE
servers:
  - url: http://localhost:8080/api/v1
    description: Local management API server
  - url: https://dns-controller.example.com/api/v1
    description: Production management API server
security:
  - ApiKeyAuth: []
...
paths:
...
  /zones:
    get:
      tags: [zones]
      summary: List all zones
      description: Returns a paginated list of zones. Agent API keys may call this operation only with fields=summary.
      operationId: listZones
      parameters:
        - name: limit
          in: query
          description: Maximum number of zones to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of zones to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: fields
          in: query
          description: Response field set. Use summary to return only zone names and versions.
          schema:
            type: string
            enum: [full, summary]
            default: full
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  zones:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/Zone'
                        - $ref: '#/components/schemas/ZoneSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [zones]
      summary: Create a new zone (JSON mode)
      description: Creates a new zone with structured JSON data
      operationId: createZone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateZoneRequest'
            examples:
              basicZone:
                summary: Basic zone with minimal records
                value:
                  name: "example.com."
                  soa:
                    mname: "ns1.example.com."
                    rname: "admin.example.com."
                    refresh: 3600
                    retry: 1800
                    expire: 604800
                    minimum: 86400
                  records:
                    - name: "@"
                      type: "NS"
                      ttl: 3600
                      value: "ns1.example.com."
                    - name: "@"
                      type: "A"
                      ttl: 300
                      value: "192.0.2.1"
      responses:
        '201':
          description: Zone created successfully
          headers:
            ETag:
              description: Zone version for optimistic concurrency control
              schema:
                type: string
            Location:
              description: URL of the created zone
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Zone already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /zones/raw:
    post:
      tags: [zones]
      summary: Create a new zone (Raw BIND mode)
      description: |
        Creates a new zone from a raw BIND zone file. Supports both text/plain and multipart/form-data formats.

        **Unsupported features:**
        - $GENERATE directive (explicitly rejected)
        - Unknown RR types including DNSSEC records (explicitly rejected)
        - $INCLUDE directive (disabled by default for security - prevents arbitrary file access)
      operationId: createZoneRaw
      parameters:
        - name: origin
          in: query
          description: Zone origin (required for text/plain without $ORIGIN directive)
          required: false
          schema:
            type: string
            example: "example.com."
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: |
                $ORIGIN example.com.
                $TTL 3600
                @   IN  SOA ns1.example.com. admin.example.com. (
                            2024122801 ; serial
                            3600       ; refresh
                            1800       ; retry
                            604800     ; expire
                            86400      ; minimum
                        )
                    IN  NS  ns1.example.com.
                    IN  NS  ns2.example.com.
                    IN  A   192.0.2.1
                www IN  A   192.0.2.2
          multipart/form-data:
            schema:
              type: object
              properties:
                zonefile:
                  type: string
                  format: binary
                  description: BIND zone file
                origin:
                  type: string
                  description: Zone origin (optional, can be auto-detected from filename or $ORIGIN)
                  example: "example.com."
              required:
                - zonefile
      responses:
        '201':
          description: Zone created successfully
          headers:
            ETag:
              description: Zone version (ETag)
              schema:
                type: string
                example: "v01ARZ3NDEKTSV4RRFFQ69G5FAV"
            Location:
              description: URL of the created zone
              schema:
                type: string
                example: "/api/v1/zones/example.com."
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: "example.com."
                  version:
                    type: string
                    example: "v01ARZ3NDEKTSV4RRFFQ69G5FAV"
                  soa:
                    type: object
                    properties:
                      serial:
                        type: integer
                        format: int64
                        example: 2024122801
                      mname:
                        type: string
                        example: "ns1.example.com."
                      rname:
                        type: string
                        example: "admin.example.com."
                      refresh:
                        type: integer
                        format: int64
                        example: 3600
                      retry:
                        type: integer
                        format: int64
                        example: 1800
                      expire:
                        type: integer
                        format: int64
                        example: 604800
                      minimum:
                        type: integer
                        format: int64
                        example: 86400
                  records_count:
                    type: integer
                    example: 4
                  message:
                    type: string
                    example: "zone successfully parsed and created from BIND format"
                  warnings:
                    type: array
                    description: Non-fatal import warnings.
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          example: "duplicate_records_deduplicated"
                        message:
                          type: string
                          example: "duplicate records were deduplicated during import"
                        count:
                          type: integer
                          example: 1
        '400':
          description: Bad request (parse error, unsupported directive, validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Zone already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /zones/{zoneName}:
    parameters:
      - name: zoneName
        in: path
        required: true
        description: Fully qualified domain name of the zone
        schema:
          type: string
          example: "example.com."

    head:
      tags: [zones]
      summary: Get zone headers
      description: Returns zone metadata headers without a response body. Requires an admin API key.
      operationId: headZone
      parameters:
        - name: If-None-Match
          in: header
          description: Zone version ETag - returns 304 if the zone matches
          schema:
            type: string
      responses:
        '200':
          description: Zone exists
          headers:
            ETag:
              description: Zone version
              schema:
                type: string
        '304':
          description: Not modified (when If-None-Match matches current version)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [zones]
      summary: Get a specific zone
      description: Returns detailed information about a zone. Requires an admin API key.
      operationId: getZone
      parameters:
        - name: If-None-Match
          in: header
          description: Zone version ETag - returns 304 if the zone matches
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          headers:
            ETag:
              description: Zone version
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '304':
          description: Not modified (when If-None-Match matches current version)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [zones]
      summary: Update a zone
      description: Updates zone SOA metadata. Existing records and DNSSEC metadata are preserved. Requires If-Match header for concurrency control.
      operationId: updateZone
      parameters:
        - name: If-Match
          in: header
          description: Zone version for optimistic concurrency control
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateZoneRequest'
      responses:
        '200':
          description: Zone updated successfully
          headers:
            ETag:
              description: New zone version
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - zone has been modified (If-Match mismatch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '428':
          description: Precondition required - If-Match header required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [zones]
      summary: Delete a zone
      description: Deletes a zone and all its records. Requires If-Match with the current zone version.
      operationId: deleteZone
      parameters:
        - name: If-Match
          in: header
          required: true
          description: Current zone ETag/version for optimistic locking
          schema:
            type: string
      responses:
        '204':
          description: Zone deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Conflict - If-Match mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          description: Precondition required - If-Match header required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Admin keys can access management APIs; agent keys are limited to zone summary listing and signed artifact reads.
    ObservabilityBearerAuth:
      type: http
      scheme: bearer
      description: Optional Bearer token for observability endpoints when observability.auth_token is configured.
  schemas:
    Zone:
      type: object
      required:
        - name
        - version
        - soa
        - records
      properties:
        name:
          type: string
          description: Fully qualified domain name
          example: "example.com."
        version:
          type: string
          description: Zone version identifier (controller-issued)
          example: "v01ARZ3NDEKTSV4RRFFQ69G5FAV"
        soa:
          $ref: '#/components/schemas/SOARecord'
        records:
          type: array
          items:
            $ref: '#/components/schemas/Record'
        dnssec:
          $ref: '#/components/schemas/DNSSECInfo'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ZoneSummary:
      type: object
      required:
        - name
        - version
      properties:
        name:
          type: string
          description: Fully qualified domain name
          example: "example.com."
        version:
          type: string
          description: Zone version identifier (controller-issued)
          example: "v01ARZ3NDEKTSV4RRFFQ69G5FAV"
    SOARecord:
      type: object
      required:
        - mname
        - rname
        - refresh
        - retry
        - expire
        - minimum
      properties:
        mname:
          type: string
          description: Primary name server
          example: "ns1.example.com."
        rname:
          type: string
          description: Admin email (with @ replaced by .)
          example: "admin.example.com."
        serial:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: Zone serial number (auto-generated)
          example: 2024122801
        refresh:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
          description: Refresh interval in seconds
          example: 3600
        retry:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
          description: Retry interval in seconds
          example: 1800
        expire:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
          description: Expiration time in seconds
          example: 604800
        minimum:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
          description: Minimum TTL in seconds
          example: 86400
    Record:
      type: object
      required:
        - name
        - type
        - ttl
        - value
      properties:
        id:
          type: string
          description: Record identifier. Backend-specific when stored; record CRUD responses derive a deterministic id when the backend does not provide one.
        name:
          type: string
          description: Record name (relative to zone)
          example: "www"
        type:
          type: string
          enum: [A, AAAA, CNAME, MX, NS, TXT, PTR, SRV, CAA]
          example: "A"
        ttl:
          type: integer
          format: int64
          minimum: 1
          maximum: 2147483647
          description: Time to live in seconds
          example: 300
        value:
          type: string
          description: Record value (format depends on type)
          example: "192.0.2.1"
        priority:
          type: integer
          format: int32
          minimum: 0
          maximum: 65535
          description: Priority (for MX/SRV records)
          nullable: true
    DNSSECInfo:
      type: object
      properties:
        enabled:
          type: boolean
        algorithm:
          type: integer
        ksk_key_tag:
          type: integer
        zsk_key_tag:
          type: integer
        nsec3_enabled:
          type: boolean
        signature_expiration:
          type: string
          format: date-time
...
    CreateZoneRequest:
      type: object
      required:
        - name
        - soa
      properties:
        name:
          type: string
        soa:
          $ref: '#/components/schemas/SOARecord'
        records:
          description: Records to create with the zone. Omit to create an empty zone.
          type: array
          items:
            $ref: '#/components/schemas/CreateRecordRequest'
    UpdateZoneRequest:
      type: object
      required:
        - name
        - soa
      properties:
        name:
          type: string
        soa:
          $ref: '#/components/schemas/SOARecord'
    CreateRecordRequest:
      type: object
      required:
        - name
        - type
        - ttl
        - value
      properties:
        name:
          type: string
        type:
          type: string
          enum: [A, AAAA, CNAME, MX, NS, TXT, PTR, SRV, CAA]
        ttl:
          type: integer
          format: int64
          minimum: 1
          maximum: 2147483647
        value:
          type: string
...
    Pagination:
      type: object
      properties:
        offset:
          type: integer
          description: Number of items skipped
        limit:
          type: integer
          description: Maximum items per page
        count:
          type: integer
          description: Number of items returned in this page
        next_offset:
          type: integer
          description: Offset to request the next page when another page may exist. Omitted when the current page is shorter than the requested limit.
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - NOT_FOUND
            - ALREADY_EXISTS
            - INVALID_INPUT
            - CONFLICT
            - INTERNAL_ERROR
            - UNAVAILABLE
            - UNAUTHORIZED
            - FORBIDDEN
            - RATE_LIMIT_EXCEEDED
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: API key role is not allowed for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'